Minor tweaks and optimizations
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@2072 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/samples/tutorial/source/main.cpp
@@ -266,7 +266,7 @@ int CompileScript(asIScriptEngine *engine)
 	// Read the entire file
 	string script;
 	script.resize(len);
-	int c =	fread(&script[0], len, 1, f);
+	size_t c = fread(&script[0], len, 1, f);
 	fclose(f);
 
 	if( c == 0 )
